+++ src/scenes/LevelSelect.ts
@@ -141,8 +141,7 @@ export default class LevelSelect extends Phaser.Scene {
 
 	private UICam!: Phaser.Cameras.Scene2D.BaseCamera | any;
 
-	private levelsKey = ['jump', 'flap', 'punch', 'airborne', 'dive', 'gunfire', 'dive-tec', 'intro', 'airborne', 'dive', 'uppercut', 'charge', 'parasol', 'pogo', 'grenade', 'final'];
-	// private levelsKey = ['intro-1', 'intro-2', 'intro-3', 'intro-4', 'intro-5', 'intro-6', 'intro-7', 'intro-8', 'intro-9', 'intro-10'];
+	private levelsKey = ['jump', 'flap', 'punch', 'airborne', 'dive', 'gunfire', 'uppercut', 'charge', 'combo', 'dive-tec', 'intro', 'parasol', 'pogo', 'grenade', 'final'];
 
 	private gamepad:Phaser.Input.Gamepad.Gamepad | undefined;
 	private SelectKey!: Phaser.Input.Keyboard.Key;